backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-playwright>=0.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0"
]

